Add empty test suite
The Travis CI configuration runs tests.  This project does not (yet)
have tests, but I would like to keep the Travis CI configuration
consistent among projects.  Stack conveniently does nothing when told to
run tests on a package with no tests, but Cabal unfortunately fails with
an error.  This commit adds an empty test suite so that Cabal will not
fail.